+++ R/group_bootstrap.R
@@ -14,13 +14,12 @@ group_bootstrap <- function(x, i, j, indent = 1, ...) {
 
 group_bootstrap_col <- function(x, j, ihead, ...) {
 
-  m <- meta(x)
+  out <- x@table_string
 
-  out <- strsplit(x, "\\n")[[1]]
+  out <- strsplit(out, "\\n")[[1]]
   header <- NULL
 
-
-  miss <- as.list(setdiff(seq_len(m$ncols), unlist(j)))
+  miss <- as.list(setdiff(seq_len(ncol(x)), unlist(j)))
   miss <- stats::setNames(miss, rep(" ", length(miss)))
   j <- c(j, miss)
 
@@ -38,44 +37,41 @@ group_bootstrap_col <- function(x, j, ihead, ...) {
 
   out <- paste(out, collapse = "\n")
 
-  class(out) <- class(x)
-  attr(out, "tinytable_meta") <- m
+  x@table_string <- out
 
-  out <- style_bootstrap(out, i = ihead, align = "c")
+  x <- style_bootstrap(x, i = ihead, align = "c")
 
   # midrule on numbered spans (not full columns of body)
   jnames <- names(j)
   jnames <- seq_along(jnames)[trimws(jnames) != ""]
-  out <- style_bootstrap(out, i = ihead, j = jnames, line = "b", line_width = .05, line_color = "#d3d8dc")
+  x <- style_bootstrap(x, i = ihead, j = jnames, line = "b", line_width = .05, line_color = "#d3d8dc")
 
-  return(out)
+  return(x)
 }
 
 
 group_bootstrap_row <- function(x, i, j, indent = 1, ...) {
   label <- names(i)
 
-  m <- meta(x)
-
   # reverse order is important
   i <- rev(sort(i))
 
+  out <- x@table_string
 
-  tab <- strsplit(x, "\\n")[[1]]
-  out <- x
+  tab <- strsplit(out, "\\n")[[1]]
 
   for (g in seq_along(i)) {
     js <- sprintf(
       "window.addEventListener('load', function () { insertSpanRow(%s, %s, '%s') });",
       # 0-indexing
-      i[g] + m$nhead - 1,
-      m$ncols,
+      i[g] + x@nhead - 1,
+      ncol(x),
       names(i)[g])
     out <- bootstrap_setting(out, new = js, component = "cell")
   }
 
   # add rows to attributes BEFORE style_tt
-  out <- meta(out, "nrows", m$nrows + length(label))
+  x@nrow <- x@nrow + length(label)
 
   # need unique function names in case there are
   # multiple tables in one Rmarkdown document
@@ -85,26 +81,28 @@ group_bootstrap_row <- function(x, i, j, indent = 1, ...) {
     out,
     fixed = TRUE)
 
-  idx <- insert_values(seq_len(m$nrows), rep(NA, length(i)), i)
+  idx <- insert_values(seq_len(nrow(x)), rep(NA, length(i)), i)
   idx_old <- idx$new[!is.na(idx$old)]
   idx_new <- idx$new[is.na(idx$old)]
 
   # limit index ot number of rows to avoid styling header or footer
-  idx_old <- idx_old[idx_old <= meta(x)$nrows]
+  idx_old <- idx_old[idx_old <= nrow(x)]
+
+  x@table_string <- out
 
   # should not be style_tt, because we already have a string bootstrap table at this stage
-  out <- style_bootstrap(out, i = idx_old, j = 1, indent = indent)
+  x <- style_bootstrap(x, i = idx_old, j = 1, indent = indent)
 
   # if there's a two-level header column multi-span, we want it centered.
-  out <- style_bootstrap(out, i = -1, align = "c")
+  x <- style_bootstrap(x, i = -1, align = "c")
 
   dots <- list(...)
   dots[["j"]] <- NULL
   if (length(dots) > 0) {
-    args <- c(list(x = out, i = idx$new[is.na(idx$old)]), dots)
-    out <- do.call(style_tt, args)
+    args <- c(list(x = x, i = idx$new[is.na(idx$old)]), dots)
+    x <- do.call(style_tt, args)
   }
 
   # do not override meta since we modified it here above
-  return(out)
+  return(x)
 }