+++ R/tt_typst.R
@@ -1,11 +1,10 @@
-tt_typst <- function(x, caption, theme, width, notes, ...) {
+tt_typst <- function(x, ...) {
   out <- readLines(system.file("templates/typst.typ", package = "tinytable"))
   out <- paste(out, collapse = "\n")
-  attr(out, "tinytable_meta") <- meta(x)
 
   # body
-  body <- apply(x, 2, function(k) paste0("[", k, "]"))
-  if (!is.null(colnames(x))) {
+  body <- apply(x@table_dataframe, 2, function(k) paste0("[", k, "]"))
+  if (!is.null(colnames(x)) && length(colnames(x)) > 0) {
     body <- rbind(paste0("[", colnames(x), "]"), body)
   }
   body <- apply(body, 1, paste, collapse = ", ")
@@ -13,29 +12,28 @@ tt_typst <- function(x, caption, theme, width, notes, ...) {
   body <- paste0(body, ",\n")
   out <- typst_insert(out, body, type = "body")
 
+  x@table_string <- out
+
   # themes ("void" needs nothing)
-  if (isTRUE(theme %in% c("default", "striped"))) {
-    out <- style_typst(out, i = 1 - meta(out, "nhead"), line = "t", line_width = .1)
-    out <- style_typst(out, i = 0, line = "b", line_width = .05)
-    out <- style_typst(out, i = meta(out)$nrows, line = "b", line_width = .1)
-  } else if (isTRUE(theme == "grid")) {
-    out <- sub("auto-lines: false,", "auto-lines: true,", out)
+  if (isTRUE(x@theme %in% c("default", "striped"))) {
+    x <- style_typst(x, i = 1 - x@nhead, line = "t", line_width = .1)
+    x <- style_typst(x, i = 0, line = "b", line_width = .05)
+    x <- style_typst(x, i = nrow(x), line = "b", line_width = .1)
+  } else if (isTRUE(x@theme == "grid")) {
+    x@table_string <- sub("auto-lines: false,", "auto-lines: true,", x@table_string)
   }
   # striped = default + background
-  if (isTRUE(theme == "striped")) {
-    out <- style_typst(out, i = seq(1, meta(out)$nrows, by = 2), background = "#ededed")
+  if (isTRUE(x@theme == "striped")) {
+    x <- style_typst(x, i = seq(1, nrow(x), by = 2), background = "#ededed")
   } 
 
-
-  class(out) <- c("tinytable", "knit_asis", class(out))
-  return(out)
+  return(x)
 }
 
 
 typst_insert <- function(x, content = NULL, type = "body") {
   if (is.null(content)) return(x)
 
-  m <- meta(x)
   out <- strsplit(x, "\n")[[1]]
   comment <- switch(type,
       "lines" = "tinytable lines before",
@@ -51,7 +49,5 @@ typst_insert <- function(x, content = NULL, type = "body") {
   }
 
   out <- paste(out, collapse = "\n")
-  class(out) <- class(x)
-  attr(out, "tinytable_meta") <- m
   return(out)
 }