+++ NEWS.md
@@ -15,6 +15,7 @@ Misc:
 * Typst now supports images and inline plots with `plot_tt()`. Thanks to @aghaynes for contribution #155.
 * Improved support for `tibble`. ANSI characters (ex: fancy `pillar` formatting) are stripped automatically or converted to HTML when the `fansi` package is installed. `fansi` is a dependency of `tibble`, so it should often be installed.
 * New `tinytable_tt_digits` global option can set the default number of digits in the `tt()` function.
+* Refactor: `tinytable` objects are now S4 class objects, with slots to hold data about the content and structure.
 
 Breaking changes: